+++ cache/page.go
@@ -13,7 +13,7 @@ var pages = make(map[string]*structs.Page) // The actual cache
 var urls = make([]string, 0)               // Duplicate of the keys in the `pages` map, but in order of being added
 var maxPages = 0                           // Max allowed number of pages in cache
 var maxSize = 0                            // Max allowed cache size in bytes
-var lock = sync.RWMutex{}
+var mu = sync.RWMutex{}
 var timeout = time.Duration(0)
 
 // SetMaxPages sets the max number of pages the cache can hold.
@@ -79,8 +79,8 @@ func AddPage(p *structs.Page) {
 		RemovePage(urls[0])
 	}
 
-	lock.Lock()
-	defer lock.Unlock()
+	mu.Lock()
+	defer mu.Unlock()
 	pages[p.URL] = p
 	// Remove the URL if it was already there, then add it to the end
 	removeURL(p.URL)
@@ -90,24 +90,24 @@ func AddPage(p *structs.Page) {
 // RemovePage will remove a page from the cache.
 // Even if the page doesn't exist there will be no error.
 func RemovePage(url string) {
-	lock.Lock()
-	defer lock.Unlock()
+	mu.Lock()
+	defer mu.Unlock()
 	delete(pages, url)
 	removeURL(url)
 }
 
 // ClearPages removes all pages from the cache.
 func ClearPages() {
-	lock.Lock()
-	defer lock.Unlock()
+	mu.Lock()
+	defer mu.Unlock()
 	pages = make(map[string]*structs.Page)
 	urls = make([]string, 0)
 }
 
 // SizePages returns the approx. current size of the cache in bytes.
 func SizePages() int {
-	lock.RLock()
-	defer lock.RUnlock()
+	mu.RLock()
+	defer mu.RUnlock()
 	n := 0
 	for _, page := range pages {
 		n += page.Size()
@@ -116,16 +116,16 @@ func SizePages() int {
 }
 
 func NumPages() int {
-	lock.RLock()
-	defer lock.RUnlock()
+	mu.RLock()
+	defer mu.RUnlock()
 	return len(pages)
 }
 
 // GetPage returns the page struct, and a bool indicating if the page was in the cache or not.
 // (nil, false) is returned if the page isn't in the cache.
 func GetPage(url string) (*structs.Page, bool) {
-	lock.RLock()
-	defer lock.RUnlock()
+	mu.RLock()
+	defer mu.RUnlock()
 
 	p, ok := pages[url]
 	if ok && (timeout == 0 || time.Since(p.MadeAt) < timeout) {