+++ config/keybindings.go
@@ -3,6 +3,7 @@ package config
 import (
 	"strings"
 
+	"code.rocketnine.space/tslocum/cview"
 	"github.com/gdamore/tcell/v2"
 	"github.com/spf13/viper"
 )
@@ -42,6 +43,10 @@ const (
 	CmdReload
 	CmdBack
 	CmdForward
+	CmdMoveUp
+	CmdMoveDown
+	CmdMoveLeft
+	CmdMoveRight
 	CmdPgup
 	CmdPgdn
 	CmdNewTab
@@ -52,6 +57,10 @@ const (
 	CmdHelp
 	CmdSub
 	CmdAddSub
+	CmdCopyPageURL
+	CmdCopyTargetURL
+	CmdBeginning
+	CmdEnd
 )
 
 type keyBinding struct {
@@ -147,35 +156,43 @@ func parseBinding(cmd Command, binding string) {
 // Called by config.Init()
 func KeyInit() {
 	configBindings := map[Command]string{
-		CmdLink1:       "keybindings.bind_link1",
-		CmdLink2:       "keybindings.bind_link2",
-		CmdLink3:       "keybindings.bind_link3",
-		CmdLink4:       "keybindings.bind_link4",
-		CmdLink5:       "keybindings.bind_link5",
-		CmdLink6:       "keybindings.bind_link6",
-		CmdLink7:       "keybindings.bind_link7",
-		CmdLink8:       "keybindings.bind_link8",
-		CmdLink9:       "keybindings.bind_link9",
-		CmdLink0:       "keybindings.bind_link0",
-		CmdBottom:      "keybindings.bind_bottom",
-		CmdEdit:        "keybindings.bind_edit",
-		CmdHome:        "keybindings.bind_home",
-		CmdBookmarks:   "keybindings.bind_bookmarks",
-		CmdAddBookmark: "keybindings.bind_add_bookmark",
-		CmdSave:        "keybindings.bind_save",
-		CmdReload:      "keybindings.bind_reload",
-		CmdBack:        "keybindings.bind_back",
-		CmdForward:     "keybindings.bind_forward",
-		CmdPgup:        "keybindings.bind_pgup",
-		CmdPgdn:        "keybindings.bind_pgdn",
-		CmdNewTab:      "keybindings.bind_new_tab",
-		CmdCloseTab:    "keybindings.bind_close_tab",
-		CmdNextTab:     "keybindings.bind_next_tab",
-		CmdPrevTab:     "keybindings.bind_prev_tab",
-		CmdQuit:        "keybindings.bind_quit",
-		CmdHelp:        "keybindings.bind_help",
-		CmdSub:         "keybindings.bind_sub",
-		CmdAddSub:      "keybindings.bind_add_sub",
+		CmdLink1:         "keybindings.bind_link1",
+		CmdLink2:         "keybindings.bind_link2",
+		CmdLink3:         "keybindings.bind_link3",
+		CmdLink4:         "keybindings.bind_link4",
+		CmdLink5:         "keybindings.bind_link5",
+		CmdLink6:         "keybindings.bind_link6",
+		CmdLink7:         "keybindings.bind_link7",
+		CmdLink8:         "keybindings.bind_link8",
+		CmdLink9:         "keybindings.bind_link9",
+		CmdLink0:         "keybindings.bind_link0",
+		CmdBottom:        "keybindings.bind_bottom",
+		CmdEdit:          "keybindings.bind_edit",
+		CmdHome:          "keybindings.bind_home",
+		CmdBookmarks:     "keybindings.bind_bookmarks",
+		CmdAddBookmark:   "keybindings.bind_add_bookmark",
+		CmdSave:          "keybindings.bind_save",
+		CmdReload:        "keybindings.bind_reload",
+		CmdBack:          "keybindings.bind_back",
+		CmdForward:       "keybindings.bind_forward",
+		CmdMoveUp:        "keybindings.bind_moveup",
+		CmdMoveDown:      "keybindings.bind_movedown",
+		CmdMoveLeft:      "keybindings.bind_moveleft",
+		CmdMoveRight:     "keybindings.bind_moveright",
+		CmdPgup:          "keybindings.bind_pgup",
+		CmdPgdn:          "keybindings.bind_pgdn",
+		CmdNewTab:        "keybindings.bind_new_tab",
+		CmdCloseTab:      "keybindings.bind_close_tab",
+		CmdNextTab:       "keybindings.bind_next_tab",
+		CmdPrevTab:       "keybindings.bind_prev_tab",
+		CmdQuit:          "keybindings.bind_quit",
+		CmdHelp:          "keybindings.bind_help",
+		CmdSub:           "keybindings.bind_sub",
+		CmdAddSub:        "keybindings.bind_add_sub",
+		CmdCopyPageURL:   "keybindings.bind_copy_page_url",
+		CmdCopyTargetURL: "keybindings.bind_copy_target_url",
+		CmdBeginning:     "keybindings.bind_beginning",
+		CmdEnd:           "keybindings.bind_end",
 	}
 	// This is split off to allow shift_numbers to override bind_tab[1-90]
 	// (This is needed for older configs so that the default bind_tab values
@@ -198,6 +215,16 @@ func KeyInit() {
 	for k, kname := range tcell.KeyNames {
 		tcellKeys[kname] = k
 	}
+
+	// Set cview navigation keys to use user-set ones
+	cview.Keys.MoveUp2 = viper.GetStringSlice(configBindings[CmdMoveUp])
+	cview.Keys.MoveDown2 = viper.GetStringSlice(configBindings[CmdMoveDown])
+	cview.Keys.MoveLeft2 = viper.GetStringSlice(configBindings[CmdMoveLeft])
+	cview.Keys.MoveRight2 = viper.GetStringSlice(configBindings[CmdMoveRight])
+	cview.Keys.MoveFirst = viper.GetStringSlice(configBindings[CmdBeginning])
+	cview.Keys.MoveFirst2 = nil
+	cview.Keys.MoveLast = viper.GetStringSlice(configBindings[CmdEnd])
+	cview.Keys.MoveLast2 = nil
 
 	for c, allb := range configBindings {
 		for _, b := range viper.GetStringSlice(allb) {