+++ display/tab.go
@@ -1,13 +1,16 @@
 package display
 
 import (
+	"fmt"
+	"net/url"
 	"strconv"
 	"strings"
 
+	"code.rocketnine.space/tslocum/cview"
+	"github.com/atotto/clipboard"
 	"github.com/gdamore/tcell/v2"
 	"github.com/makeworld-the-better-one/amfora/config"
 	"github.com/makeworld-the-better-one/amfora/structs"
-	"gitlab.com/tslocum/cview"
 )
 
 type tabMode int
@@ -122,15 +125,100 @@ func makeNewTab() *tab {
 		}
 	})
 	t.view.SetInputCapture(func(event *tcell.EventKey) *tcell.EventKey {
-		// Capture left/right scrolling and change the left margin size accordingly, see #197
-		// Up/down scrolling is saved in this func to keep them in sync, but the keys
-		// are passed and no extra behaviour happens.
+		// Capture scrolling and change the left margin size accordingly, see #197
+		// This was also touched by #222
+		// This also captures any tab-specific events now
+
+		if t.mode != tabModeDone {
+			// Any events that should be caught when the tab is loading is handled in display.go
+			return nil
+		}
+
+		cmd := config.TranslateKeyEvent(event)
+
+		// Cmds that aren't single row/column scrolling
+		//nolint:exhaustive
+		switch cmd {
+		case config.CmdBookmarks:
+			Bookmarks(&t)
+			t.addToHistory("about:bookmarks")
+			return nil
+		case config.CmdAddBookmark:
+			go addBookmark()
+			return nil
+		case config.CmdPgup:
+			t.pageUp()
+			return nil
+		case config.CmdPgdn:
+			t.pageDown()
+			return nil
+		case config.CmdSave:
+			if t.hasContent() {
+				savePath, err := downloadPage(t.page)
+				if err != nil {
+					Error("Download Error", fmt.Sprintf("Error saving page content: %v", err))
+				} else {
+					Info(fmt.Sprintf("Page content saved to %s. ", savePath))
+				}
+			} else {
+				Info("The current page has no content, so it couldn't be downloaded.")
+			}
+			return nil
+		case config.CmdBack:
+			histBack(&t)
+			return nil
+		case config.CmdForward:
+			histForward(&t)
+			return nil
+		case config.CmdSub:
+			Subscriptions(&t, "about:subscriptions")
+			tabs[curTab].addToHistory("about:subscriptions")
+			return nil
+		case config.CmdCopyPageURL:
+			currentURL := tabs[curTab].page.URL
+			err := clipboard.WriteAll(currentURL)
+			if err != nil {
+				Error("Copy Error", err.Error())
+				return nil
+			}
+			return nil
+		case config.CmdCopyTargetURL:
+			currentURL := t.page.URL
+			selectedURL := t.HighlightedURL()
+			if selectedURL == "" {
+				return nil
+			}
+			u, _ := url.Parse(currentURL)
+			copiedURL, err := u.Parse(selectedURL)
+			if err != nil {
+				err := clipboard.WriteAll(selectedURL)
+				if err != nil {
+					Error("Copy Error", err.Error())
+					return nil
+				}
+				return nil
+			}
+			err = clipboard.WriteAll(copiedURL.String())
+			if err != nil {
+				Error("Copy Error", err.Error())
+				return nil
+			}
+			return nil
+		}
+		// Number key: 1-9, 0, LINK1-LINK10
+		if cmd >= config.CmdLink1 && cmd <= config.CmdLink0 {
+			if int(cmd) <= len(t.page.Links) {
+				// It's a valid link number
+				followLink(&t, t.page.URL, t.page.Links[cmd-1])
+				return nil
+			}
+		}
+
+		// Scrolling stuff
 
 		key := event.Key()
 		mod := event.Modifiers()
-		ru := event.Rune()
-
-		width, height := t.view.GetBufferSize()
+		height, width := t.view.GetBufferSize()
 		_, _, boxW, boxH := t.view.GetInnerRect()
 
 		// Make boxW accurate by subtracting one if a scrollbar is covering the last
@@ -140,8 +228,7 @@ func makeNewTab() *tab {
 			boxW--
 		}
 
-		if (key == tcell.KeyRight && mod == tcell.ModNone) ||
-			(key == tcell.KeyRune && mod == tcell.ModNone && ru == 'l') {
+		if cmd == config.CmdMoveRight || (key == tcell.KeyRight && mod == tcell.ModNone) {
 			// Scrolling to the right
 
 			if t.page.Column >= leftMargin() {
@@ -158,28 +245,37 @@ func makeNewTab() *tab {
 				}
 			}
 			t.page.Column++
-		} else if (key == tcell.KeyLeft && mod == tcell.ModNone) ||
-			(key == tcell.KeyRune && mod == tcell.ModNone && ru == 'h') {
+		} else if cmd == config.CmdMoveLeft || (key == tcell.KeyLeft && mod == tcell.ModNone) {
 			// Scrolling to the left
 			if t.page.Column == 0 {
 				// Can't scroll to the left anymore
 				return nil
 			}
 			t.page.Column--
-		} else if (key == tcell.KeyUp && mod == tcell.ModNone) ||
-			(key == tcell.KeyRune && mod == tcell.ModNone && ru == 'k') {
+		} else if cmd == config.CmdMoveUp || (key == tcell.KeyUp && mod == tcell.ModNone) {
 			// Scrolling up
 			if t.page.Row > 0 {
 				t.page.Row--
 			}
 			return event
-		} else if (key == tcell.KeyDown && mod == tcell.ModNone) ||
-			(key == tcell.KeyRune && mod == tcell.ModNone && ru == 'j') {
+		} else if cmd == config.CmdMoveDown || (key == tcell.KeyDown && mod == tcell.ModNone) {
 			// Scrolling down
 			if t.page.Row < height {
 				t.page.Row++
 			}
 			return event
+		} else if cmd == config.CmdBeginning {
+			t.page.Row = 0
+			// This is required because cview will also set the column (incorrectly)
+			// if it handles this event itself
+			t.applyScroll()
+			App.Draw()
+			return nil
+		} else if cmd == config.CmdEnd {
+			t.page.Row = height
+			t.applyScroll()
+			App.Draw()
+			return nil
 		} else {
 			// Some other key, stop processing it
 			return event
@@ -207,18 +303,27 @@ func (t *tab) addToHistory(u string) {
 
 // pageUp scrolls up 75% of the height of the terminal, like Bombadillo.
 func (t *tab) pageUp() {
-	row, col := t.view.GetScrollOffset()
-	t.view.ScrollTo(row-(termH/4)*3, col)
+	t.page.Row -= (termH / 4) * 3
+	if t.page.Row < 0 {
+		t.page.Row = 0
+	}
+	t.applyScroll()
 }
 
 // pageDown scrolls down 75% of the height of the terminal, like Bombadillo.
 func (t *tab) pageDown() {
-	row, col := t.view.GetScrollOffset()
-	t.view.ScrollTo(row+(termH/4)*3, col)
+	height, _ := t.view.GetBufferSize()
+
+	t.page.Row += (termH / 4) * 3
+	if t.page.Row > height {
+		t.page.Row = height
+	}
+
+	t.applyScroll()
 }
 
 // hasContent returns false when the tab's page is malformed,
-// has no content or URL, or if it's an 'about:' page.
+// has no content or URL.
 func (t *tab) hasContent() bool {
 	if t.page == nil || t.view == nil {
 		return false
@@ -226,13 +331,15 @@ func (t *tab) hasContent() bool {
 	if t.page.URL == "" {
 		return false
 	}
-	if strings.HasPrefix(t.page.URL, "about:") {
-		return false
-	}
 	if t.page.Content == "" {
 		return false
 	}
 	return true
+}
+
+// isAnAboutPage returns true when the tab's page is an about page
+func (t *tab) isAnAboutPage() bool {
+	return strings.HasPrefix(t.page.URL, "about:")
 }
 
 // applyHorizontalScroll handles horizontal scroll logic including left margin resizing,
@@ -322,4 +429,16 @@ func (t *tab) applyAll() {
 	if t == tabs[curTab] {
 		t.applyBottomBar()
 	}
+}
+
+// HighlightedURL returns the currently selected URL
+func (t *tab) HighlightedURL() string {
+	currentSelection := tabs[curTab].view.GetHighlights()
+
+	if len(currentSelection) > 0 {
+		linkN, _ := strconv.Atoi(currentSelection[0])
+		selectedURL := tabs[curTab].page.Links[linkN]
+		return selectedURL
+	}
+	return ""
 }