+++ display/util.go
@@ -5,9 +5,9 @@ import (
 	"net/url"
 	"strings"
 
+	"code.rocketnine.space/tslocum/cview"
 	"github.com/makeworld-the-better-one/go-gemini"
 	"github.com/spf13/viper"
-	"gitlab.com/tslocum/cview"
 	"golang.org/x/text/unicode/norm"
 )
 
@@ -90,7 +90,7 @@ func textWidth() int {
 // It also returns an error if it could not resolve the links, which should be displayed
 // to the user.
 func resolveRelLink(t *tab, prev, next string) (string, error) {
-	if !t.hasContent() {
+	if !t.hasContent() || t.isAnAboutPage() {
 		return next, nil
 	}